Drop the unneeded wrapper of  PhotoCollage and extending the array of Photo instead for the configure function
