+++ LessLinesDemo/Datasource/TableViewCellType.swift
@@ -10,7 +10,7 @@ import UIKit
 
 enum TableViewCellType {
     case photoItem(Photo)
-    case photoCollage(PhotoCollage)
+    case photoCollage([Photo])
     case spacing
 }
 
@@ -19,8 +19,8 @@ extension TableViewCellType {
         switch self {
         case .photoItem(let photo):
             return TableViewCellDescriptor(reuseIdentifier: ItemTableViewCell.className, configure: photo.configure)
-        case .photoCollage(let photoCollage):
-            return TableViewCellDescriptor(reuseIdentifier: PhotoItemsCollageTableViewCell.className, configure: photoCollage.configure)
+        case .photoCollage(let photosArray):
+            return TableViewCellDescriptor(reuseIdentifier: PhotoItemsCollageTableViewCell.className, configure: photosArray.configure)
         case .spacing:
             return TableViewCellDescriptor(reuseIdentifier: SpacingTableViewCell.className, configure: { _ in })
         }
