--- LessLinesDemo/Models/PhotoCollage.swift
@@ -1,13 +1,0 @@
-//
-//  PhotoCollage.swift
-//  LessLinesDemo
-//
-//  Created by Mustafa Baalbaki on 11/21/18.
-//  Copyright © 2018 Mustafa. All rights reserved.
-//
-
-import Foundation
-
-struct PhotoCollage {
-    var photos: [Photo]
-}
