+++ LessLinesDemo/ViewModels/Photo+ViewModel.swift
@@ -20,3 +20,21 @@ extension Photo {
     }
 }
 
+extension Array where Element == Photo {
+    func configure(_ cell: PhotoItemsCollageTableViewCell) {
+        let items = DatasourceFactory.dataSourceForPhotoItemsCollage(self)
+        let flowLayout = UICollectionViewFlowLayout()
+        flowLayout.minimumLineSpacing = 7
+        flowLayout.minimumInteritemSpacing = 7
+        flowLayout.scrollDirection = .vertical
+        let aspectRatio = CGFloat(375.0/524.0)
+        let collectionViewLayoutFactory = CollectionViewLayoutFactory(flowLayout: flowLayout,
+                                                                      phonePortraitItemsCount: 2,
+                                                                      phoneLandscapeItemsCount: 4,
+                                                                      padPortraitItemsCount: 7,
+                                                                      padLandscapeItemsCount: 12,
+                                                                      itemAspectRatio: aspectRatio)
+        cell.configure(collectionViewLayoutFactory: collectionViewLayoutFactory, items: items)
+    }
+}
+
